% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spatial_outlier.R
\name{find_spatial_outlier}
\alias{find_spatial_outlier}
\title{Generate two dimensional feature space from tensor data.}
\usage{
find_spatial_outlier(
  data,
  ncomp = 2,
  center = TRUE,
  center.mode = c("A", "B", "C"),
  scale = TRUE,
  scale.mode = c("A", "B", "C"),
  alpha = 0.01,
  tensor_decom = c("Tucker3", "parafac")
)
}
\arguments{
\item{data}{3-way array of data, A mode- space}

\item{ncomp}{Number of components. Default is set to 2}

\item{center}{Whether to center the data}

\item{center.mode}{If centering the data, on which mode to do this}

\item{scale}{Whether to scale the data}

\item{scale.mode}{scaling the data, on which mode to do this}

\item{alpha}{Threshold for determining the cutoff for outliers.}

\item{tensor_decom}{Method to decompose the 3-way array}
}
\value{
a list with components
\item{outliers}{The indexes of the spatial observations determined to be outliers.}
\item{out_scores}{Outliying score of each spatial observation}
\item{type}{Type of each spatial observation: outlier or typical}
\item{plot_data}{A-mode component data with ourlying measures}
}
\description{
This function converts the original three way dataset into
two dimensional feature space.
}
\examples{
library(rrcov3way)
data(Arno)
result <- find_spatial_outlier(
  data = Arno, tensor_decom = "Tucker3"
)
result$out_data
}
