% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_mask}
\alias{generate_mask}
\title{Create a masking function}
\usage{
generate_mask(envir, hidden)
}
\arguments{
\item{envir}{An environment that will be used to mask others.}

\item{hidden}{A character vector of package names.}
}
\value{
A masking function.
}
\description{
Create a masking function
}
