% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Evaluate a function or expression in a masking environment}
\usage{
mask(x, envir = rlang::caller_env())
}
\arguments{
\item{x}{A function or an expression}

\item{envir}{An environment or a name of a package as a `character`, which
should be injected between the execution environment and the current
parent environment.}
}
\value{
In the case that `x` is a function, the result is a a copy of
  the function in a new enclosing environment.  Otherwise, the result of
  `x` executing in the new environment.
}
\description{
Evaluate a function or expression in a masking environment
}
\examples{
mask(mtcars \%>\% dplyr::select(mpg, cyl), "tidylog")
}
