% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_outliers.R
\name{plot_spatial_outliers}
\alias{plot_spatial_outliers}
\title{Display spatial outliers in two dimensional A-mode component plot}
\usage{
plot_spatial_outliers(X, label_outlier = TRUE, nudge_y = 0.05)
}
\arguments{
\item{X}{A-mode component data with ourlying measures produced by \code{\link{find_spatial_outlier}}}

\item{label_outlier}{Whether to display labels of outlying observations. Default is set to TRUE.}

\item{nudge_y}{Vertical adjustment to nudge labels by. Useful for offsetting text from points, particularly on discrete scales.}
}
\value{
A ggplot object of data space with detected outliers (if any).
}
\description{
Provide a 2D scatterplot of data for visual exploration.
For data with more than two dimensions, two dimensional scatterplot is produced
using the first two pricipal components.
}
\examples{
library(rrcov3way)
data(Arno)
result <- find_spatial_outlier(
  data = Arno, center.mode = "B",
  scale.mode = "B", tensor_decom = "Tucker3"
)
p <- plot_spatial_outliers(X = result$out_data)
p + viridis::scale_color_viridis(discrete = TRUE) +
ggplot2::xlim(-1.1,0.1)

}
