% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{tidylog_mask}
\alias{tidylog_mask}
\title{Mask with tidylog}
\usage{
tidylog_mask(x)
}
\arguments{
\item{x}{A function or expression.}
}
\value{
If x evaluates to a function, then a new function is returned
  which is masked by the tidylog namespace.  Otherwise, x is evaluated
  in a child environment of the tidylog namespace.
}
\description{
Make use of tidylog without masking in the global environment.
}
