% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\docType{package}
\name{mask}
\alias{mask}
\title{mask: Anomaly Detection in Multivariate Spatio-temporal data with K-measurements}
\description{
This package provides a framework to detect spatial anomalies in
 multivariate spatio-temporal data (tensor data, multiway data).
 An anomaly is a spatial point or region that deviates significantly
 from the global and/or local distribution of a given network.
}
\seealso{
The core functions in this package:  \code{\link{find_spatial_outlier}},
\code{\link{plot_spatial_outliers}}
}
