% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined.R
\name{tidylog_mask}
\alias{tidylog_mask}
\title{Create a mask for tidylog evaluation}
\usage{
tidylog_mask()
}
\arguments{
\item{hidden}{A character vector of package names that should be in}
}
\value{
A masking function.
}
\description{
Create a mask for tidylog evaluation
}
\examples{
m <- tidylog_mask()
m( select(mtcars, mpg))
}
